requirements.txt
quart==0.18.1
line-bot-sdk==3.1.0
aiohttp==3.8.4
hypercorn
werkzeug==2.3.0
